+++ src/components/marquee/index.tsx
@@ -1,0 +1,3 @@
+export const Marquee: React.FC = () => {
+    return <div></div>;
+};