+++ src/components/section/index.tsx
@@ -12,10 +12,7 @@ export const Section: React.FC<Props> = ({
     className = "",
 }: Props) => {
     return (
-        <section
-            id={title}
-            className={`flex flex-col w-full ${className}`}
-        >
+        <section id={title} className={`flex flex-col w-full ${className}`}>
             {children}
         </section>
     );