+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [0.3.0] - 2022-01-28
+### Added
+- All Posts page (unstyled)
+- Read Post page (unstyled)
+- Notion API utilities to read Blog posts from Notion Database
+- Tailwind Typography plugin for post rendering
+- Blog components
+- Post cache in JSON format
+- Highlight.js for code highlighting
+### Removed
+- Boilerplate example files
+- Zustland
+### Fixed
+- Header alignment
+- Image filters
+- React.FC replaced by NextPage in pages
+
 ### [0.2.2] - 2022-01-28
 ### Added
 - Button components